fix: separate backend from web app and made a new app only for backend
