+++ apps/backend/next.config.ts
@@ -1,0 +1,8 @@
+import type { NextConfig } from "next";
+
+const nextConfig: NextConfig = {
+  transpilePackages: ["@fulbito/utils", "@fulbito/types"],
+};
+
+export default nextConfig;
+
