+++ apps/backend/src/app/api/auth/[...nextauth]/authOptions.ts
@@ -1,0 +1,46 @@
+import type { NextAuthOptions } from 'next-auth'
+import Credentials from 'next-auth/providers/credentials'
+import { prisma } from '@/lib/prisma'
+import bcrypt from 'bcryptjs'
+
+export const authOptions: NextAuthOptions = {
+  secret: process.env.NEXTAUTH_SECRET,
+  debug: process.env.NODE_ENV === 'development',
+  session: { strategy: 'jwt' },
+  providers: [
+    Credentials({
+      name: 'Credentials',
+      credentials: {
+        username: { label: 'Username', type: 'text' },
+        password: { label: 'Password', type: 'password' }
+      },
+      async authorize(credentials) {
+        if (!credentials?.username || !credentials?.password) return null
+        const user = await prisma.user.findUnique({ where: { username: credentials.username } })
+        if (!user) return null
+        const ok = await bcrypt.compare(credentials.password, user.passwordHash)
+        if (!ok) return null
+        return { id: user.id, name: user.username, role: user.role }
+      }
+    })
+  ],
+  callbacks: {
+    async jwt({ token, user }) {
+      if (user) {
+        token.role = ((user as unknown as { role?: string }).role === 'ADMIN' ? 'ADMIN' : 'USER')
+      }
+      return token
+    },
+    async session({ session, token }) {
+      if (session.user) {
+        ;(session.user as unknown as { role?: string }).role = (token as unknown as { role?: string }).role || 'USER'
+        if (token.sub) session.user.id = token.sub
+      }
+      return session
+    }
+  },
+  pages: {
+    signIn: '/login'
+  }
+}
+
