+++ apps/backend/src/lib/auth.ts
@@ -1,5 +1,5 @@
 import { getServerSession } from 'next-auth'
-import { authOptions } from '@/lib/authOptions'
+import { authOptions } from '@/app/api/auth/[...nextauth]/authOptions'
 
 export async function getSession() {
   return getServerSession(authOptions)
@@ -12,5 +12,4 @@ export async function requireAdmin(): Promise<{ ok: true } | { ok: false, status
   if (role !== 'ADMIN') return { ok: false, status: 403, body: { error: 'Forbidden' } }
   return { ok: true }
 }
-
