+++ apps/backend/src/types/next-auth.d.ts
@@ -18,4 +18,3 @@ declare module 'next-auth/jwt' {
 
 export {}
 
-
