+++ apps/web/package.json
@@ -3,10 +3,9 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "postinstall": "prisma generate",
     "dev": "next dev",
-    "build": "prisma generate && next build",
-    "build:production": "prisma migrate deploy && prisma generate && next build",
+    "build": "next build",
+    "build:production": "next build",
     "start": "next start",
     "lint": "eslint"
   },
@@ -14,12 +13,9 @@
     "@fulbito/types": "workspace:*",
     "@fulbito/utils": "workspace:*",
     "@heroicons/react": "^2.2.0",
-    "@prisma/client": "^6.15.0",
-    "bcryptjs": "^3.0.2",
     "next": "15.5.9",
     "next-auth": "^4.24.12",
     "nodemailer": "^7.0.10",
-    "prisma": "^6.15.0",
     "react": "19.1.0",
     "react-dom": "19.1.0",
     "react-icons": "^5.5.0",
