--- apps/web/prisma/schema.prisma
@@ -1,64 +1,0 @@
-generator client {
-  provider = "prisma-client-js"
-}
-
-datasource db {
-  provider = "postgresql"
-  url      = env("DATABASE_URL")
-}
-
-model Player {
-  id               String        @id @default(cuid())
-  name             String
-  position         String
-  skill            Float?
-  skills           Json?
-  createdAt        DateTime      @default(now())
-  updatedAt        DateTime      @updatedAt
-  photoUrl         String?
-  shirtDutiesCount Int           @default(0)
-  shirtMatches     Match[]       @relation("ShirtDutyMatches")
-  matchPlayers     MatchPlayer[]
-}
-
-model Match {
-  id                  String        @id @default(cuid())
-  date                DateTime
-  type                String
-  name                String?
-  teamAScore          Int
-  teamBScore          Int
-  createdAt           DateTime      @default(now())
-  updatedAt           DateTime      @updatedAt
-  shirtsResponsibleId String?
-  shirtsResponsible   Player?       @relation("ShirtDutyMatches", fields: [shirtsResponsibleId], references: [id])
-  players             MatchPlayer[]
-}
-
-model MatchPlayer {
-  id          String @id @default(cuid())
-  matchId     String
-  playerId    String
-  team        String
-  goals       Int
-  performance Float
-  match       Match  @relation(fields: [matchId], references: [id], onDelete: Cascade)
-  player      Player @relation(fields: [playerId], references: [id], onDelete: Cascade)
-
-  @@index([playerId])
-  @@index([matchId])
-}
-
-model User {
-  id           String   @id @default(cuid())
-  username     String   @unique
-  passwordHash String
-  role         Role     @default(USER)
-  createdAt    DateTime @default(now())
-  updatedAt    DateTime @updatedAt
-}
-
-enum Role {
-  USER
-  ADMIN
-}
