--- apps/web/src/app/api/auth/[...nextauth]/route.ts
@@ -1,9 +1,0 @@
-export const runtime = 'nodejs'
-import NextAuth from 'next-auth'
-import { authOptions } from './authOptions'
-
-const handler = NextAuth(authOptions)
-
-export { handler as GET, handler as POST }
-
-
