--- apps/web/src/app/api/upload/route.ts
@@ -1,25 +1,0 @@
-export const runtime = 'nodejs'
-import { NextResponse } from 'next/server'
-import { requireAdmin } from '@/lib/auth'
-
-export async function POST(req: Request) {
-  const gate = await requireAdmin()
-  if (!gate.ok) return NextResponse.json(gate.body, { status: gate.status })
-  try {
-    const form = await req.formData()
-    const file = form.get('file') as File | null
-    if (!file) return NextResponse.json({ error: 'No file uploaded' }, { status: 400 })
-
-    const arrayBuffer = await file.arrayBuffer()
-    const buffer = Buffer.from(arrayBuffer)
-
-    const mime = (file as unknown as { type?: string }).type || 'image/png'
-    const base64 = buffer.toString('base64')
-    const dataUrl = `data:${mime};base64,${base64}`
-    return NextResponse.json({ url: dataUrl })
-  } catch {
-    return NextResponse.json({ error: 'Upload failed' }, { status: 500 })
-  }
-}
-
-
