+++ apps/web/src/app/players/playersClient.tsx
@@ -5,8 +5,9 @@ import { useSession } from 'next-auth/react'
 import SkillBadge from '@/components/SkillBadge'
 import StreakBadge from '@/components/StreakBadge'
 import { calculateAllCurrentStreaks } from '@/lib/playerStats'
-import { useMatchStore, Match } from '@/store/useMatchStore'
-import { usePlayerStore, Player } from '@/store/usePlayerStore'
+import type { Match, Player } from '@fulbito/types'
+import { useMatchStore } from '@/store/useMatchStore'
+import { usePlayerStore } from '@/store/usePlayerStore'
 import { useEffect, useMemo, useState, useCallback, useRef } from 'react'
 import Image from 'next/image'
