+++ apps/web/src/lib/backend.ts
@@ -1,0 +1,4 @@
+export function getBackendBaseUrl() {
+  return process.env.BACKEND_URL || 'http://localhost:3001'
+}
+
