--- apps/web/src/lib/prisma.ts
@@ -1,11 +1,0 @@
-import { PrismaClient } from '@prisma/client'
-
-const globalForPrisma = global as unknown as { prisma?: PrismaClient }
-
-export const prisma =
-  globalForPrisma.prisma ??
-  new PrismaClient({
-    log: process.env.NODE_ENV === 'development' ? ['query', 'error', 'warn'] : ['error']
-  })
-
-if (process.env.NODE_ENV !== 'production') globalForPrisma.prisma = prisma
