+++ apps/web/src/lib/shape.ts
@@ -1,22 +1,21 @@
-import { prisma } from '@/lib/prisma'
-import type { Player as StorePlayer } from '@/store/usePlayerStore'
-import type { Match as StoreMatch } from '@/store/useMatchStore'
+import type { Match, Player } from '@fulbito/types'
+import { getBackendBaseUrl } from '@/lib/backend'
 
-export function shapeStorePlayers(players: Array<{ id: string; name: string; position: string; skill: number | null; skills?: unknown; photoUrl?: string | null; createdAt: Date; updatedAt: Date }>): StorePlayer[] {
+export function shapeStorePlayers(players: Array<{ id: string; name: string; position: string; skill: number | null; skills?: unknown; photoUrl?: string | null; createdAt: Date | string; updatedAt: Date | string }>): Player[] {
   return players.map((p) => ({
     id: p.id,
     name: p.name,
     position: p.position,
     skill: p.skill ?? null,
-    skills: p.skills as StorePlayer['skills'],
+    skills: p.skills as Player['skills'],
     photoUrl: p.photoUrl ?? undefined,
     shirtDutiesCount: (p as unknown as { shirtDutiesCount?: number }).shirtDutiesCount ?? 0,
-    createdAt: p.createdAt,
-    updatedAt: p.updatedAt,
+    createdAt: p.createdAt instanceof Date ? p.createdAt : new Date(p.createdAt),
+    updatedAt: p.updatedAt instanceof Date ? p.updatedAt : new Date(p.updatedAt),
   }))
 }
 
-export function shapeStoreMatches(matches: Array<{ id: string; date: Date; type: string; name: string | null; teamAScore: number; teamBScore: number; shirtsResponsibleId?: string | null; players: Array<{ playerId: string; team: 'A' | 'B'; goals: number; performance: number }> }>, nameMap: Map<string, string>): StoreMatch[] {
+export function shapeStoreMatches(matches: Array<{ id: string; date: Date; type: string; name: string | null; teamAScore: number; teamBScore: number; shirtsResponsibleId?: string | null; players: Array<{ playerId: string; team: 'A' | 'B'; goals: number; performance: number }> }>, nameMap: Map<string, string>): Match[] {
   return matches.map((m) => ({
     id: m.id,
     date: m.date.toISOString().slice(0, 10),
@@ -34,13 +33,16 @@ export function shapeStoreMatches(matches: Array<{ id: string; date: Date; type:
   }))
 }
 
-export async function loadPlayersAndMatches(): Promise<{ players: StorePlayer[]; matches: StoreMatch[] }> {
-  const [players, matches] = await Promise.all([
-    prisma.player.findMany({ orderBy: { name: 'asc' } }),
-    prisma.match.findMany({ orderBy: { date: 'desc' }, include: { players: true } }),
+export async function loadPlayersAndMatches(): Promise<{ players: Player[]; matches: Match[] }> {
+  const base = getBackendBaseUrl()
+  const [playersRes, matchesRes] = await Promise.all([
+    fetch(`${base}/api/players`, { cache: 'no-store' }),
+    fetch(`${base}/api/matches`, { cache: 'no-store' }),
   ])
-  const nameMap = new Map(players.map((p) => [p.id, p.name] as const))
+  const players = await playersRes.json()
+  const matches = await matchesRes.json()
+
+  // `GET /api/matches` already returns store-shaped matches (with names)
   const shapedPlayers = shapeStorePlayers(players)
-  const shapedMatches = shapeStoreMatches(matches.map(m => ({ ...m, players: m.players.map(mp => ({ playerId: mp.playerId, team: mp.team as 'A' | 'B', goals: mp.goals, performance: mp.performance })) })), nameMap)
-  return { players: shapedPlayers, matches: shapedMatches }
+  return { players: shapedPlayers, matches }
 }
